requirements.txt
requests
beautifulsoup4
pyyaml
pandas
streamlit
Pillow
python-dotenv